requirements.txt
pandas
sqlalchemy 
psycopg2-binary
python-dotenv
fastapi
uvicorn
httpx
flask
flask_cors
requests
scikit-learn
requests 
beautifulsoup4
matplotlib
google.genai
duckduckgo-search
google-search-results
ddgs
bcrypt